Move all Bluetooth-related functionalities from MainActivity's ViewModel to BluetoothPage's
