+++ app/src/main/java/ecb/manifest/kowalski/obd_scan/ui/main/MainActivityViewModel.kt
@@ -11,27 +11,4 @@ import kotlinx.coroutines.flow.combine
 import kotlinx.coroutines.flow.stateIn
 import javax.inject.Inject
 
-@HiltViewModel
-class MainActivityViewModel @Inject constructor(
-    private val bluetoothController: IBluetoothController
-) : ViewModel() {
-    private val _state = MutableStateFlow(BluetoothUiState())
-    val state = combine(
-        bluetoothController.scannedDevices,
-        bluetoothController.pairedDevices,
-        _state
-    ) { scannedDevices, pairedDevices, state ->
-        state.copy(
-            scannedDevices = scannedDevices,
-            pairedDevices = pairedDevices,
-        )
-    }.stateIn(viewModelScope, SharingStarted.WhileSubscribed(5000), _state.value)
-
-    fun startScan() {
-        bluetoothController.startDiscovery()
-    }
-
-    fun stopScan() {
-        bluetoothController.stopDiscovery()
-    }
-}
+class MainActivityViewModel : ViewModel() { }
